# api/various/swagger_output.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "This API provides information about pets and checkups for a veterinarian office",
      "description": ""
    },
    "host": "nevercricket-viennahazard-3000.codio.io",
    "basePath": "/",
    "schemes": [
      "https"
    ],
    "paths": {
      "/pets": {
        "get": {
          "description": "The GET action on this endpoint provides a list of all the pets clients, it takes no parameters.",
          "parameters": [],
          "responses": {}
        },
        "post": {
          "description": "he POST action on this endpoint allows you to add a pet. The required parameters are ‘kind’ and 'name’.",
          "parameters": [
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/pets/{petId}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "put": {
          "description": "",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "id": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/pets/{petId}/checkups": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {}
        }
      },
      "/pets/{petId}/checkups/{checkupId}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "checkupId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "checkupId",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {}
        }
      }
    }
  }